+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [0.4.5-preview.2] - 2020-10-27
+## This is the v0.4.5-preview.1 release of *Unity Game Simulation*
+- add metric names to editor window
+
 ## [0.4.5-preview.1] - 2020-09-21
 ## This is the v0.4.5-preview.1 release of *Unity Game Simulation*
 - fix api parameters